Table: row order may not be updated because lengths of the arrays differ
If a row is moved and then deleted in the same request, the row order
changed event contains too less rows. The UI will throw an exception if
the arrays differ.

This may be solved by using getTableRowId(row) != null instead of
isRowAccepted in handleModelRowOrderChanged. Because isRowAccepted
returns false if the row was deleted, but the row order change event
needs that row.

Even though this would work, I could create a similar case which cannot
be fixed like that (see JsonTableTest). That is why we decided to append
the missing rows to the row order changed event. Since they are deleted
anyway it is fine.

The only other solution we could think of was to always move the row
order changed event to the end of all events and update it with the
current state of rows. Then the event would always contain the correct
amount of rows.

216635

Change-Id: I42ab641c4fb2cb91467ddb579d87adf8dfb9e519
Reviewed-on: https://git.eclipse.org/r/110931
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>