Show own page menus on detail form
If a page-with-nodes contains a detail form, the menus from the page are
copied to the detail form so that the action can directly be performed
in the form.
The menus to be copied must be the ones from the page on which the
detail form is registered (not the ones currently active on the
outline). Otherwise the menus point to wrong elements.

#216234

Change-Id: I7393ff2b2514adc20950a4bb8ed80bded133e21d
(cherry picked from commit 947d49d22cab94b2bcc33dff4528eb5b07887a37)
Reviewed-on: https://git.eclipse.org/r/110977
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>