FormField: support horizontal alignment for Scout JS
GridData.horizontalAlignment is not only used to align the widget in the
grid cell, but sometimes also for the inner alignment of the field.

This did not work for Scout JS apps, because _renderGridData is not
called when the gridData changes, it is only used if the grid is
calculated on on server side. To make it work, _renderGridDataHints has
been added because the developer should always work with the
gridDataHints rather than the gridData directly.