Popup: prevent JS error when opening and closing many popups very fast
When the "popupOpen" event on the desktop is fired, other popups could
be closed. However, those other popups may still be called for the same
event, even though they have detached their listener in the _remove()
method! This is because scout.events always creastes a copy of the
current listener list before executing the callbacks. Therefore, the
_onPopupOpen handler has to check if the popup is still rendered before
accessing the this.$container property.

Change-Id: I46c250925151d15bb0627971cdf3c18037376ecf
Reviewed-on: https://git.eclipse.org/r/111203
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/111234