Table: remove cell editor popup on detach (fix JS error in FocusContext)
Because unlike all other popups, CellEditorPopup is not closed by the UI
but by the server (when "endCellEditor" request has been completed) the
cell editor popup may still be open when the table widget is detached.
This could cause errors in the FocusContext._focus() method, because the
new element to focus does not belong to the active focus context.

How to reproduce the error:
- Open two forms with an editable table.
- Click in the table so that the cell editor popup is opened.
- While the popup is open, click the form tab of the other form.
- JS error happens. 

This is a backport of commit 3b223006b72ae07f8bf4717f72dd69cdd9e1b9c5
from branch 6.1.x.

216732

Change-Id: I8330914409f6a016512ef1b357e42dd368044662
Reviewed-on: https://git.eclipse.org/r/111202
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>