HtmlComponent: add widthHint support
The widthHint may be used to determine the preferredSize of a component
based on a given width.

The accordion needs this feature because every group is as big as its
content. If the content are tiles, the height depends on the width
because the tiles will wrap on smaller widths.