Bug 527658 Allow usage of JmsMomImplementor in J2EE environments
According to specification not all API methods of JMS are allowed to be
used in a J2EE environment.
Also Session, MessageProducer and MessageConsumer are not thread safe
objects. Therefore these objects should not be shared.

Add Context credentials to create connection method.

In case of single threaded subscription, now the ISubscription#dispose()
method waits for any ongoing processing of this subscription to finish.

J2EE constraints supported:
 - Each Session requires its own Connection
 - ClientId should not be set on Connection
 - ExceptionListener should not be set on Connection
 - Async send (CompletionListener) is not supported
 - Async receive (MessageListener) is not supported

Not supported / implementation dependent:
 - Set ClientId on ConnectionFactory
 - Transactions in J2EE / JTA

https://bugs.eclipse.org/bugs/show_bug.cgi?id=527658

Change-Id: I504bb1b4dc79f5aa4651adf7f6c1cd6ef696db74
Signed-off-by: Ralph Steiner <steiner.ralph@gmail.com>
Reviewed-on: https://git.eclipse.org/r/111699
Tested-by: Hudson CI
Reviewed-by: Ralph Steiner <rst@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/113230
Reviewed-on: https://git.eclipse.org/r/113242
Tested-by: Patrick B??nziger <patrick.baenziger@bsi-software.com>