Fix wrong size of tiles in an accordion
If the accordion has tiles with uiHeight = true and an active filter,
the height of the tiles is likely to be wrong after the filter is
removed and the tiles shown.

Reason: prefSize is called during the animation which runs when the tile is made visible. Because the animation scales the tile, the height is different to the height after the animation.