Widget: don't accept layout operations while removing
Use case: Removing the tree while the collapse animation is still
running. 
This will fail because it tries to revalidate the layout on each step.
Because _remove of the tree does not just stop the animation but let it
jump to the end (to make the complete function execute), a revalidation
will be executed while removing the tree. And in that case an exception
occured because the the menu bar had already been removed.

217627