Close all open proposal choosers when UiSession is started
Avoid that proposal choosers from content assist fields are still opened
in the Java model after the UI has been reloaded by pressing Ctrl + R.
This is only required for the 'old' smart-field, the new smart-field has
no model representation of the proposal chooser anymore. That means when
the UI is reloaded, the proposal chooser is gone. Thus, this method
mimics the behavior of the new smart-field. 

216067