Fix tree selection issues after using enter to navigate
When pressing enter in a detail table to select a node, the node won't
be selected and the default detail form is shown instead. After that
even selecting the node with the mouse is node possible anymore.

This is because filterAccepted of the node is set to false (by the
LazyNodeFilter). The state is actually wrong because the node should be
visible but in order to do that we would have to keep track of the
clicked nodes. Because this requires too many changes in the tree we
decided to partially revert the commit which broke the behavior (which
is Table/Tree: prevent programmatically selecting filtered rows/nodes,
3ec1792e25680b702b32f68381f994c903b7db93).

216779