Tree: make area for node control bigger
The node control was quite hard to click and even harder to touch with
the finger because it is very small.

To fix it some padding was added around the control. It is now also
positioned with padding-left which means the area left of the control
toggles the expansion on click, too.

Further changes:
- The icon itself is now 1px bigger on a regular tree to be the same
height as the text. The outline control icon still has the same size.
- The node padding-left may now be set by CSS without having to adjust
the JS code.
- The CSS variable @tree-node-padding has been splitted into
@tree-node-padding-y, @tree-node-padding-left and
@tree-node-padding-right.