ImageField: don't remove other image source properties if one is set
This functionality has been added because the UI may only read one image
source property. But actually it is not necessary to delete the other
properties because there is a defined order which the UI considers when
reading the properties. It even breaks existing code because one might
rely on that behavior. Example: imageId and image are set. If the image
is removed, the imageId will be shown instead.