Widget: add postInitConfig, init and dispose
Goal is to harmonize all the init methods (initField, initTile, initForm
etc.) and to make sure, init is only executed once. This is important
for dynamic widgets like accordion or tiles. These widgets initialize
the newly added children by themselves so that the caller does not need
to take care of it. For these cases it is important that init() is not
called multiple times (see AccordionTest for details).

But: there may be cases which require init() to be called multiple
times, like reseting a search form. For such cases, reinit() has to be
used from now on.

Also moved initConfig guard (isInitConfigDone) to AbstractWidget so that
individual widgets don't have to care about it and to use the same
pattern as for postInitConfig, init and dispose.

Change-Id: Idf33876b08b43b6887d401b96d6b1a004854bfb2