GzipServletFilter: enable compression based on content types
The existing implementation enabled the compression based on the file
ending. This does not work for typical REST API calls because they don't
have a file ending.

The new implementation checks the content type of the response. The file
ending check is removed.

Also added some more content types compared to file endings. This means
some more content will be compressed from now on which is xml, svg and
vcard.

218966