TilesAdapter: ignore 'selectedTiles' events while synching tiles
When the 'tiles' property is updated from the server, the existing tiles
are removed. This causes an existing tile selection to be set to []
(empty). However, when using Scout classic, the list of selected tiles
is maintained by the UI server. Therefore, any updates of the
'selectedTiles' property that are simply caused as a side-effect of
setting the new tiles list should be ignored. Otherwise, the server-side
'selectedTiles' property will be overridden with [].

Change-Id: Ib548650ab3df64f58d131c35e5d0089924cfb2aa