Configure comment formatter to use old-style comment length.
The source code formatter in Eclipse before Oxygen counts the comment
length from the start of the comment marker. Starting with Oxygen this
default behavior was changed to count from the first character instead.
This has some effect on existing comments: more text can fit on one
line. But this also changes all javadoc comments which leads to changes
in the file even though nothing was actually edited. Therefore this new
feature was disabled to have the same line length counting strategy as
in older Eclipse versions. See Bug 514017 for details.