Don't process navigate button key shortcuts when form is open
Now the buttons register their keystroke on the keystroke-context of the
table or the form (previously the outline has been used). With this
solution the keystroke still works when the outline is collapsed but not
when a form is open.

Note: there's a small hack by using a different property for the
keyStrokeContext, which was required because of the init/register order
of the keystrokes/-contexts. It would be better if we could simply use
the _createKeyStrokeContext function, however this would cause too many
changes, so I've preferred this simple solution.

215646