AbstractForm: use isSaveNeeded instead of own collector in doCancel
doCancel checked if there is any field that has isSaveNeeded set to
true. If so, the corresponding message box was shown (depending on the
value of isAskIfNeedSave).

doOk and doSave check isSaveNeeded instead, which is calculated be the
call to checkSaveNeeded. This calculation starts at the root group box
and recursively traverses its children. That way, a group box might
override execIsSaveNeeded so that its children are no relevant anymore
(i.e. by just returning false). The visitor used by doCancel did not
take this into account. Thus it was possible that doCancel showed the
message box but doStore did not store because isSaveNeeded was false.

Change-Id: Ib92d25b8ca200c6209dff11beb4fd10e848270f6
Reviewed-on: https://git.eclipse.org/r/113308
Tested-by: Hudson CI
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>
Reviewed-by: Andreas Hoegger <Andreas.Hoegger@bsiag.com>