TilesAccordion: implement correct selection behavior and add JS version
Selection behavior in a tiles accordion affects the accordion as a whole
and not each tile grid individually. This means the UI has to know that
there are other tile grids when a tile is selected, especially while
pressing the ctrl key. In order to implement that, TilesAccordion.js has
been introduced. This gives the possibility to handle the selection in a
different way for the accordion than for a single tile grid.

Adding a TilesAccordion.js not only fixes the selection behavior but
also adds the possibility to use the widget for Scout JS apps.

Because the UI side TilesAccordion now processes the same logic as the
java one, it requires a synchronization with the UI of some properties
(selectable, multiSelect), otherwise the UI would deselect tiles even
though multiSelect is enabled on server side.