Set proper cache headers on login and logout responses
/login and /logout requests are forwarded to /login.html resp.
/logout.html. This forwarding prevented the setting of correct cache
control headers which means these pages are never cached.
Furthermore from a security point of view missing cache control headers
on a logout response are bad because the outcome will be undefined. It
is important that the logout request will be sent to the server so that
the server session may be invalidated properly -> the headers
must-revalidate and maxAge=0 have to be present.

219023

Change-Id: I674849da9763b134dee7ab84b96b4b7e97b7711d
Reviewed-on: https://git.eclipse.org/r/113643
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>