Selenium: try to improve test stability by waiting for server calls
When the network connection is very slow, tests sometimes fail because
they try to access an element that is not (yet) fully rendered on
screen, because the corresponding server call has not been completed.

In other causes, the target element is visible but cannot be clicked,
because of a glasspane that is caused by a pending server call.

A third case is a model class with multiple instances. When the selector
is executed too early, the "old" instance may be returned because the
server call that changes the form or outline content has not been
completed. This will result in "element is stale" exceptions when the
element is accessed after the server call has finally completed.

This commit tries to improve test stability by adding more
"waitUntilDataRequestPendingDone()" calls to critical tests.

Change-Id: Ieaa367104f523ab821f7426f892e10cde5cd864c