DynamicResourceLoader: Fix download errors in IE11 for certain filenames
IE 11 cannot download files if the filename hint set in the
Content-Disposition header (both filename= and filename*=) contains
certain control characters.

This strips out all control characters in the range 0x00-0x1f.
Characters in the range 0x80-0x9f as well as 0x7f can be parsed by IE 11
and do not need to be removed.

Note that the control characters in ISO-8859-1 have the same values in
the Java UTF-16 string representation, thus removing them from the Java
string will also ensure that they will not be present in the final
header.

This also adds testcases for the current implementation.

213904

Change-Id: Iadf5ff413ab13f0b60fdfdeb6a5ec9a62e77ff82
Signed-off-by: Jeremie Miserez <jeremie.miserez@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/114820
Tested-by: Hudson CI
Reviewed-by: Andre Wegmueller <awe@bsiag.com>