TileGrid: tile grid is broken when removed while tile removal is running
The deletion of a tile triggers a remove animation. If the tile grid is
removed while the animation is running, the variable tileRemovalPending
is never reset and therefore every layout update will be prevented.

Fixed Widget.js so that remove is executed if the parent is removed even
though the remove animation has not started yet or is still running.