FocusManager: don't activate a context covered by a glasspane
This fixes a bug with IE:
A modal dialog on a modal dialog is open which has a focused text area.
Clicking in another field will activate the form which will activate the
other dialog first because the other dialog is its display parent.
Activating that dialog will activate its focus context and therefore
tries to focus the last focused field. This is not possible because it
is covered by a glass pane and therefore the desktop will be focused
(even though I am wondering if that is really correct). This has the
effect that it is not possible anymore to leave the focused text area.

223285

Based on commit: File Chooser: prevent focus loss after choosing a file
(506aad85810b491fef9cbdc55a0a0730e95531be)

Change-Id: I18e29746b6db034ec1a23eeaedce0ed1646c8fbd
Signed-off-by: Christoph Bräunlich <christoph.braeunlich@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/115051
Reviewed-by: Andre Wegmueller <awe@bsiag.com>
Tested-by: Hudson CI