Fixed Tree bug in case of autoCheckChildNodes=true
The issue occurred in a TreeBox where autoCheckChildNodes was set to
true. When the user clicked on a node with child nodes, the node and all
its children were checked correctly. But when uncheckAllKeys() was
called only the parent node was unchecked, but not the child nodes,
which is wrong. This was introduced with an earlier bug fix (213620).

With this change we send the state of all child nodes of a parent node.

218516