Tree: Do not call _rebuildParent if no parent node exists (breadcrumb)
Previously, attempting to programmatically select a non-expanded node
that is currently not visible (because another node is selected) in the
breadcrumb style tree view would lead to a error (cannot read property
'expanded' of undefined), if the selected node was a top-level node
without a parent. _rebuildParent() ensures that the node's parent nodes
are also made visible when a node is selected. It is only necessary if a
parent exists, and is now only called if that is the case.

210280

Change-Id: Ic1df785dee3e383789c189d505bbb016e870c7f4
Signed-off-by: Jeremie Miserez <jeremie.miserez@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/115261
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>