Table: Do not add an additional checkable column if one is configured.
This prevents an additional checkable column from being added when
getConfiguredCheckableColumn is set. This bug was previously fixed but
reintroduced with the fix for Bug 528935.

- Added Jasmine test for Bug 528935 and for this change.

- Added Javadoc: Changing the configured checkable column at runtime has
no effect after the table configuration is initialized.

- Note about toggling the table checkable flag at runtime: While
currently possible, it is not recommended to check/uncheck rows in the
model while the table has checkable false. Any such changes will not be
reflected in the UI. Check/Uncheck in the UI is not possible while
checkable is false.

- Updated JsonColumn to set the value of the configured checkable column 
regardless of whether the table is set to checkable during initial 
configuration. This enables loading a table with checkable initially 
set to false and still use a configured checkable column instead of an 
added one.

Change-Id: I963cda8fe59b329a5459450d5a548babc9b2d94d
Signed-off-by: Jeremie Miserez <jeremie.miserez@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/115550
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>