JsonPropertyChangeEvent: ensure property order
Properties should be sent in the same order as inserted because the ui
might rely on that -> use LinkedHashMap rather than a regular HashMap.

This is not a fix for a concrete issue because it has not made any
problems so far. But this might change because of a more intensive usage
of properties (e.g. by the tile grid).