TooltipSupport: install() should update existing tooltip
When a tooltip support is already installed on a component and
scout.tooltips.install() is called again, the existing tooltip support
should be updated accordingly. Otherwise, changing the tooltip text of a
widget using the tooltip support (e.g. a menu) would not have an effect.

The function scout.tooltips.update() can also be used manually to update
the text of an already showing tooltip. (This is especially useful when
using the same tooltip support for multiple components and retrieving
the text from the corresponding data attribute.)

A new method scout.tooltips.close() was added to manually close an
already showing tooltip.

This change includes some jasmine tests.

Change-Id: I315982672729a00c2f2f43c97b30da16b16662b2
Reviewed-on: https://git.eclipse.org/r/115678
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>