Event buffers should be @Beans to make them replaceable
If an application defines new model events (e.g. TableEvent), it should
be possible to change the corresponding event buffer logic accordingly.
This was not easily possible, because each model instance created its
own event buffer instance in a createEventBuffer() method. Because this
method is not included in the extension chains, it would be very
cumbersome to override this method on each use of the model class.

Instead of creating a new instance of an event buffer using the "new"
keyboard, the new instances are now created via BeanManager. This allows
to replace the implementation globally.

Change-Id: If4664f992d2c4190c24dcd24692a066844f6f2e6
Reviewed-on: https://git.eclipse.org/r/116260
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>