FormField: add alternative field style
The alternative style is the new default. Nevertheless, in some
circumstances is the classic style preferable, e.g. when used in a cell
editor or on a form with background color like the search form. That is
why the style is set to classic automatically if the field is used in a
cell editor or on a table control form (e.g. search form).

Also
- improved disabled style on various fields
- added hover color for clear icon
- adjusted active style for alternative file inputs
- adjusted ellipsis icon for form fields to make it better visible
- fixed menu items height to align them with form fields in menu bar
when zoomed
- made sure CompositeField does setDisabledStyle and setFieldStyle
recursively
- increased mandatory width to move the indicator a little more away
from the field