Button: show ellipsis if label is too long
It is not possible to set overflow: hidden on the button itself because
the focus border is drawn outside. It is possible to set overflow:
hidden on the label but that requires a fixed width.

Added ButtonLayout which sets a max-width in the same way the
RadioButtonLayout already does.

RadioButtonLayout now extends ButtonLayout to make prefSize work as
expected (return the correct prefSize value even if ellipsis are shown).

Added tooltip which is shown if text is truncated.

To make the ellipsis tooltip work with IE the pointer-events hack had to
be removed. This means that the active state of link buttons does not
work with IE, but this is not used by the default scout theme anyway.

215523