SmartField: fix resetDisplayText() II
Because _formatValue() may return a promise instead of a string, the
return value must not be assigned to the displayText property
immediately. Otherwise, _renderDisplayText() might throw an error,
because a promise is not a string.

Fix to the previous fix (f76c9badfec5f95ad85aaf53b745e6ca6a7e321c):
The existing method _updateDisplayText() cannot be used here, because it
internally calls setDisplayText(). This function does nothing, when the
current property value is equal to the new value. However, there are
cases where the text visible to the user does not match the
"displayText" properties (e.g. while typing). In that case, the display
text would not really be reset when leaving the field, because the
internal state would still be the same.
Instead, a custom implementation was used for resetDisplayText() which
uses the new internal _setAndRenderDisplayText() method. That method is
similar to setDisplayText(), but does not check for equality of the
displayText property.

Change-Id: I529c91705182fc64204dbf9702e363c43766964c
Reviewed-on: https://git.eclipse.org/r/116257
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>