Improve check box and radio button style
- Added new less variables for styling.
- Aligned checkbox undefined state, checkbox has-children state and
radio button checked state.
- Colorized radio button bullet the same way as check box marker.
- Used line-height to vertically align checkbox marker and (more
importantly) to make it independent of the used font-family (because the
browser may use a different line-height for different font families).

Fixing the position of the check box before element was necessary
because otherwise the check box or label jumped when getting clicked in
Chrome with 125% zoom and the new 1px border.