Style: make environment properties configurable using css
The properties of scout.HtmlEnvironment define some global sizes like
the default logical grid row height or the default field label width.
Since some of this properties are available as less variables too, one
had to adjust the values in JavaScript and CSS code. To make it easier,
the HtmlEnvironment now reads the values from CSS so that it is now
sufficient to adjust the LESS variables in sizes.less.