String field with masked input: prevent sending of display text to UI layer
A string field with masked input did send the value from the model to
the UI layer allowing to read possibly sensitive data via JS.
The value is not sent to the UI layer anymore resulting in a slightly
different behavior. From the model perspective nothing changed.

* A `value` property change in the model will send the property
`input_obfuscated` with `true` and an obfuscated display text
(`----------`).
* On focus of the input field, the obfuscated text is removed, showing
an empty input field. On blur, if nothing was typed, the obfuscated text
is restored (model value is unchanged).
* When text is typed into the input field, `input_obfuscated` property
is set to `false` and the display text represents the real display text
(but still masked).
* Clicking the clear icon will remove the display text and
`input_obfuscated` flag will be set to `false`. The clear icon is
visible in case of obfuscation even if there is no text  displayed
because field is focused.
* A `input_masked` property change in the model to `false` will send the
real display text to replace the obfuscated display text.
* A `input_masked` property change in the model to `true` will not
trigger an obfuscation of the current display text.
* When insert text is used and the display text is obfuscated, the text
to insert is the new display text (i.e. current display text is replaced
by text to insert).
* When insert text is used and the display text is not obfuscated, a
regular insert text is executed (i.e. text is added to current display
text).
* When the input field has the focus and a value is set via model, the
display text is cleared because field has focus and input is obfuscated.
When leaving the field, obfuscated text will be restored.

219037

Change-Id: I02438d3362e77b0534e441bf843268d34d8e18ae
Reviewed-on: https://git.eclipse.org/r/116528
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/116696
Reviewed-by: Stephan Merkli <stephan.merkli@bsi-software.com>