SimpleTabArea: don't wrap tab items
It may happen in Firefox that tabs will wrap if exact measuring is
enabled (tested by setting exact to true in graphics and removed
rounding from jquery-scout width/height). Exact measuring is still
disabled but may be enabled in the future -> wrapping can be disabled in
any case.

Also fixed a wrong variable access in MenuBarLayout.