Added possibility to clear global cache via URL parameter
Only allowed in development mode. This is useful for development and
running Selenium tests. Because you can run the server in cache=true
mode, which means you don't have to wait for scripts and less files to
be built on each request. However, on the first request the server is
notified to clear its cache so each run of of a test or a test suite has
the latest version of the scripts. As a developer you may add the
?clearCache=true parameter when needed and work with cached scripts most
of the time.