Session: don't close busy request on cancel request
When the busy indicator is showing and the user clicks the "cancel"
button, a "json?cancel" request is sent to the server. The busy
indicator is then changed to a canceling state, but should still be
showing. It must not be closed until the original JSON request has
finished. To achieve this behavior, sending the ?cancel request should
not trigger the busy indicator itself.

221223

Change-Id: Iac0c31785e6b91947c0c63fc592809a3929482a9
Reviewed-on: https://git.eclipse.org/r/117023
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>