Session: don't close busy request on cancel request
When the busy indicator is showing and the user clicks the "cancel"
button, a "json?cancel" request is sent to the server. The busy
indicator is then changed to a canceling state, but should still be
showing. It must not be closed until the original JSON request has
finished. To achieve this behavior, sending the ?cancel request should
not trigger the busy indicator itself.

221223

Change-Id: I6fe799ad326ae2257ca2f2eefd9df2bb81e6321b
Reviewed-on: https://git.eclipse.org/r/117023
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>