Planner: improve minSelectionInterval
When last hour of day is not midnight, clicking in the last interval
currently draws the selection only around that last interval, even
though the actual selection range is bigger. To not confuse the user
clicking in the last interval now moves the start of the selection to
the left until minSelectionInterval is reached. This is the same
behavior as when selecting the last interval of the view range.

217660