Jasmine: add expect().nothing() to remove warnings on certain tests
Jasmine tests that have no expectations (but check for exceptions) cause
warnings in the console. To suppress those warnings, a dummy expectation
for "nothing" was added.

This is a preliminary fix for the following jasmine issue:
https://github.com/jasmine/jasmine/issues/1221

When we update the jasmine test, the custom matcher in JasmineScout.js
can be removed again (see FIXME marker).

Change-Id: Icb83479994d0ed40d03e3a66bf34f1550e95b822
Reviewed-on: https://git.eclipse.org/r/117153
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>