Automatic Preloading of Web Fonts
To support theme-specific font preloading, an auto-detection feature was
added to the font preloader. It is still possible to manually a static
array of font names to preload. To use the auto-detection feature,
remote the 'fonts' property from the bootstrap object passed to
App.init(). To prevent preloading of any fonts, set the font option to
an empty array [].

Change-Id: I5fa486a7308aff9bee5b2aacbfb8f2a4011bb876