RunContexts.copyCurrent adds cancellable which is never unregistered
org.eclipse.scout.rt.platform.context.RunContexts.copyCurrent()
registered a new cancellable with the current run monitor which is never
unregistered. This is not a problem if the current run monitor finishes
after a reasonable amount of time, however it may be problematic if the
current run monitor is used in an infinite loop (e.g. client
notification poller).

This commit therefore changes the behavior for the run context to not
register the cancellable anymore at the point of copy. It adds a new
"parent run monitor" to the run context. At the start of execution the
new run monitor is added to the parent run monitor as a cancellable and
it is removed immediately after finishing execution.

222822

Conflicts:
	org.eclipse.scout.rt.client/src/main/java/org/eclipse/scout/rt/client/clientnotification/ClientNotificationPoller.java

Change-Id: Idc5db97143caa5f9f1ca606dcaea80c3b77e0fb9
Reviewed-on: https://git.eclipse.org/r/117654
Reviewed-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>
Tested-by: Matthias Otterbach <matthias.otterbach@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/118528
Tested-by: Hudson CI