Improve disabled style
- Global disabled color reverted to a darker one
  Reason: fields which only contain text (html fields, label fields
etc.) should use a readable color. The user cannot interact with such
fields anyway so he does not really need to know whether it is enabled
or not.
- Use darker color for labels again
  Reason: It gives a cleaner look and it is not necessary since the
input itself already uses the lighter color
- Aligned disabled colors for tree and table
  A disabled row / node uses the lighter color, no background color.
  A disabled table / tree uses the darker color.
- Fixed disabled colors for checkboxes in table
  Checkbox checked and undefined element used the wrong color sometimes
(e.g. if only the table-row was disabled)
- Fixed disabled and read-only color for multi line smart field in
alternative mode
- Removed rule to make mandatory indicator disappear when field is
disabled
  The rule is broken since 6.1 and had no effect anyway -> mandatory
indicator is shown even if field is disabled