FormField: improve accessibility
Linked label with the input by using aria-labelledby which allows screen
readers to read the label if an input is focused.

Also added a click listener to the labels so that clicking a label will
activate the field. This is especially helpful on mobile devices when
the new alternative style is active, because the field boundaries are
not obvious anymore.