FieldStyle: use hard edges only with alternative style
Classic fields should still have rounded borders. It looks better if
form fields with classic style are used in a menu bar near a button
which has rounded borders as well. Another advantage is that classic
style does not change to the previous release.

But it means that it is not suggested to mix the 2 styles on a form.
Otherwise you get some controls with edged and some with rounded
borders.