Table: fix auto optimize width measurement for cells with images
The auto optimize width calculation of a column needs the images to be
loaded, otherwise the width will be wrong because non loaded images have
a width of 0. The images are loaded asynchronously unless they are
loaded from cache. This means the calculation has to be delayed until
all images are loaded. The measurement div must not be removed until the
images are loaded otherwise the image load event would never be
triggered.

In order to get a correct result ALL images have to be loaded so if the
table has a lot of different (non cached images), auto resizing a column
may lead to a lot of requests.

214326

Change-Id: I19c7b423c1f6cff568bf6d9278d6357d17f3c290
Reviewed-on: https://git.eclipse.org/r/118286
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>
Tested-by: Hudson CI