Make sure selected proposal is stored when form is closed
Without the bugfix a user could type something into the smart-field, the
proposal chooser is opened, a proposal is selected, but then the user
clicks directly on the OK button. In that case the already selected
proposal was ignored, and the form was stored with an unexpected value.

Note that this bugfix does not solve the case where the lookup is still
in progress (but not yet resolved) and the user clicks OK. In that case
the form is still stored, without the results of the lookup. This case
is harder to solve, since we have no way to (blocking) wait until the
lookup is done before we save/close the form.

221944