Bugfix for SmartField2 issue with execValidateValue and exceptions
SmartField2 did not handle exceptions thrown in execValidateValue
properly. One issue was that the error was removed by the UI as soon as
the user clicked somewhere. Another issue was that the lookup row was
set on the Java model, even though the value was not set because
execValidateValue did throw an exception.

Introduced new method in ValueFieldAdapter.js to sort properties. Also
use this method in DateFieldAdapter.js, replaces slightly flawed
previous implementation.

222979