Fix mobile transformation
With the new generic widget initialization the transformation broke.
Reason: the form is now initialized before the fields. The
transformation checked if the form had already been transformed which
was always true for fields.

Since it is now possible to add fields even if the form is already open,
the check has to be removed anyway.

Also, the grid data needs to be rebuild directly due to the same
reasons. But I don't think it is necessary to rebuild every grid data,
it should be sufficient to rebuild the parent. Actually it should not
really be necessary to rebuild the whole grid, it should be sufficient
to only adjust the gridData so that the UI will be informed because
weightY/X or useUiHeight don't affect the model grid.