Table: Ensure setTableChanging counter decremented
- It is necessary to ensure that m_tableChanging is always correctly
decremented, even if Errors or Exceptions are thrown. Initially, this
was implemented by catching all Throwables and executing the rest of the
method before rethrowing the first occuring error at the end. This
worked, but was later refactored to only catch Exceptions instead of all
Throwables. Thus, PlatformErrors such as e.g. ThreadInterruptedError
were not caught, leading to invalid counter states (always > 0) if the
user cancelled at the right time. This could lead to OOM due to
unbounded growth of the event buffers m_eventBuffer and the
m_rowValueChangeBuffer.

- This commit ensures that the counter is always decremented by using
finally-blocks to make it clear that the code must always run. This is
similar to the the other set...Changing methods (in AbstractTable,
AbstractCalendar).

222277

Change-Id: If75d1acf3bed05b43b31e5c4516f689f5e295be6
Signed-off-by: Jeremie Miserez <jeremie.miserez@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/118436
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>