TileGrid: add tileClick and tileAction events
- Add tileClick and tileAction events for TileGrid, TileAccordion
- Add exec methods to TileAccordion

Difference to table:
- The tileAction is fired always even if a non selectable tile is
clicked
- The tileAction is fired always even if execTileAction is overridden
- There is no default menu because you can just call menu.doClick() in
execTileAction. This means 'Enter' is not automatically linked to
execRowAction resp. the default menu. That code in AbstractTable is
actually legacy code. You should always define the enter key stroke on
the menu so that it can be visualized when pressing F1. Furthermore the
check for isMethodOverride does not work with extensions anyway.