Copy to clipboard does not work with iOS
This seems to be a WebKit bug:
https://bugs.webkit.org/show_bug.cgi?id=176980.

The bug is set as resolved but it looks like Safari does not use the
latest WebKit yet.

The current behavior is as follows: copy button is visible, user clicks
copy. Pasting is not possible because the paste button is not visible.

To workaround it, event.preventDefault() must not be called, which means
custom plain text data cannot be set and the default copy behavior of
Safari has to be used, which copies rich text.

216581